% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchTheme.R
\name{configureLightTheme}
\alias{configureLightTheme}
\title{Configure current theme as light theme}
\usage{
configureLightTheme()
}
\value{

}
\description{
Uses an environment variable RSTUDIO_LIGHT_THEME
}
