% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchTheme.R
\name{getEditorTheme}
\alias{getEditorTheme}
\title{Get current editor theme}
\usage{
getEditorTheme()
}
\value{

}
\description{
Use this to find your current theme
and get help on setting the
the environment variables
RSTUDIO_DARK_THEME and RSTUDIO_LIGHT_THEME
}
\details{
Run it twice - once in dark mode, once with
a light theme - and follow the instructions
in the console.
}
