% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchTheme.R
\name{switchTheme}
\alias{switchTheme}
\title{switchTheme}
\usage{
switchTheme()
}
\description{
This funcions checks if the current theme is `Tomorrow Night`. If this
is the case, it changes the theme to the default theme (`Textmate`).
Likewise, the theme is changed to the dark one if it wasn't previously.
}
\examples{
# Use the addins browser or activate via the addins menue to switch between themes.

}
