% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchTheme.R
\name{switchTheme}
\alias{switchTheme}
\title{switchTheme}
\usage{
switchTheme()
}
\description{
This funcions checks if the current theme is dark. If this
is the case, it changes to the light variant.
Likewise, the theme is changed to the dark one if it wasn't previously.
}
\details{
You need the environment variables RSTUDIO_LIGHT_THEME
and RSTUDIO_DARK_THEME in your .Renviron file.
Run `usethis::edit_r_environ()` to edit it.
}
\examples{
# Use the addins browser or activate via the addins menue to switch between themes.

}
